# static/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Bus Booking System API",
      "description": "API for managing bus bookings, customers, drivers, buses, schedules, and admins.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://127.0.0.1:5555/",
        "description": "Development server"
      }
    ],
    "paths": {
      "/customers": {
        "post": {
          "summary": "Create a new customer",
          "description": "Registers a new customer in the system.",
          "operationId": "createCustomer",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "firstname": {
                      "type": "string",
                      "example": "John"
                    },
                    "lastname": {
                      "type": "string",
                      "example": "Doe"
                    },
                    "email": {
                      "type": "string",
                      "example": "john.doe@example.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "password123"
                    },
                    "address": {
                      "type": "string",
                      "example": "123 Main St"
                    },
                    "phone_number": {
                      "type": "string",
                      "example": "1234567890"
                    },
                    "id_or_passport": {
                      "type": "string",
                      "example": "123456789"
                    }
                  },
                  "required": ["firstname", "lastname", "email", "password", "address", "phone_number", "id_or_passport"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Customer successfully created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Invalid input"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/buses": {
        "post": {
          "summary": "Add a new bus",
          "description": "Registers a new bus in the system.",
          "operationId": "createBus",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string",
                      "example": "bus_123"
                    },
                    "driver_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "cost_per_seat": {
                      "type": "integer",
                      "example": 50
                    },
                    "number_of_seats": {
                      "type": "integer",
                      "example": 30
                    },
                    "route": {
                      "type": "string",
                      "example": "City Center - Airport"
                    },
                    "travel_time": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-08-09T10:00:00Z"
                    },
                    "number_plate": {
                      "type": "string",
                      "example": "ABC1234"
                    }
                  },
                  "required": ["username", "driver_id", "cost_per_seat", "number_of_seats", "route", "travel_time", "number_plate"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Bus successfully created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Bus"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Invalid input"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/drivers": {
        "post": {
          "summary": "Add a new driver",
          "description": "Registers a new driver in the system.",
          "operationId": "createDriver",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Jane Doe"
                    },
                    "license_number": {
                      "type": "string",
                      "example": "DL123456"
                    },
                    "phone_number": {
                      "type": "string",
                      "example": "0987654321"
                    },
                    "email": {
                      "type": "string",
                      "example": "jane.doe@example.com"
                    }
                  },
                  "required": ["name", "license_number", "phone_number", "email"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Driver successfully created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Driver"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Invalid input"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/drivers/{id}": {
        "get": {
          "summary": "Get a driver",
          "description": "Retrieves a driver by ID.",
          "operationId": "getDriver",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Driver details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Driver"
                  }
                }
              }
            },
            "404": {
              "description": "Driver not found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Driver not found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Customer": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "firstname": {
              "type": "string",
              "example": "John"
            },
            "lastname": {
              "type": "string",
              "example": "Doe"
            },
            "email": {
              "type": "string",
              "example": "john.doe@example.com"
            },
            "address": {
              "type": "string",
              "example": "123 Main St"
            },
            "phone_number": {
              "type": "string",
              "example": "1234567890"
            },
            "id_or_passport": {
              "type": "string",
              "example": "123456789"
            },
            "role" :{
              "type":"string",
              "example":"customer"

            }
          }
        },
        "Bus": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "username": {
              "type": "string",
              "example": "bus_123"
            },
            "driver_id": {
              "type": "integer",
              "example": 1
            },
            "cost_per_seat": {
              "type": "integer",
              "example": 50
            },
            "number_of_seats": {
              "type": "integer",
              "example": 30
            },
            "route": {
              "type": "string",
              "example": "City Center - Airport"
            },
            "travel_time": {
              "type": "string",
              "format": "date-time",
              "example": "2024-08-09T10:00:00Z"
            },
            "number_plate": {
              "type": "string",
              "example": "ABC1234"
            }
          }
        },
...
        "Driver": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "firstname": {
              "type": "string",
              "example": "Jane"
            },
            "lastname": {
              "type": "string",
              "example": "Doe"
            },
            "license_number": {
              "type": "string",
              "example": "DL123456"
            },
            "phone_number": {
              "type": "string",
              "example": "0987654321"
            },
            "email": {
              "type": "string",
              "example": "jane.doe@example.com"
            },
            "role" :{
              "type":"string",
              "example":"customer"

            }
          }
        },
...
      }
    }
  }